refactor : cookie 관련 한군데로 모으기
