+++ backend/src/main/java/com/daily/daily/auth/controller/AuthController.java
@@ -2,15 +2,13 @@ package com.daily.daily.auth.controller;
 
 import com.daily.daily.auth.dto.LoginDTO;
 import com.daily.daily.auth.dto.TokenDTO;
-import com.daily.daily.auth.jwt.JwtUtil;
 import com.daily.daily.auth.jwt.RefreshToken;
 import com.daily.daily.auth.service.AuthService;
 import com.daily.daily.auth.service.TokenService;
 import com.daily.daily.common.dto.SuccessResponseDTO;
+import com.daily.daily.common.service.CookieService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -19,13 +17,13 @@ import org.springframework.web.bind.annotation.*;
 public class AuthController {
     private final AuthService authService;
     private final TokenService tokenService;
-    private final JwtUtil jwtUtil;
+    private final CookieService cookieService;
 
     @PostMapping("/login")
     public SuccessResponseDTO login(@RequestBody LoginDTO loginDto, HttpServletResponse response) {
         TokenDTO tokenDto = authService.login(loginDto);
 
-        jwtUtil.setTokensInCookie(response, tokenDto.getAccessToken(), tokenDto.getRefreshToken());
+        cookieService.setTokensInCookie(response, tokenDto.getAccessToken(), tokenDto.getRefreshToken());
 
         return new SuccessResponseDTO();
     }
